requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
python-dateutil>=2.8.2
msgpack>=1.0.5
aiosqlite>=0.19.0  # For SQLite async support
websockets>=12.0  # For WebSocket support
python-dotenv>=1.0.0  # For environment variable management

# Logging dependencies
python-json-logger>=2.0.7  # For structured JSON logging
structlog>=23.2.0  # For structured logging

# Metrics dependencies
prometheus-client>=0.19.0  # For Prometheus metrics
prometheus-fastapi-instrumentator>=6.1.0  # FastAPI Prometheus integration

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0  # For coverage reporting
aiohttp>=3.9.0  # For concurrent request testing
locust>=2.20.0  # For load testing
pytest-benchmark>=4.0.0  # For performance benchmarking
pytest-xdist>=3.5.0  # For parallel test execution
